Update dynamic Greeting
I really liked your personal website and decided to use it as a first project in Vue. I used Nuxt to build the project so I had to adapt some things. I was curious about why the first greeting was not dynamic, so I tried to find a fix and found that you can actually have it dynamic.